% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{auto.colSet}
\alias{auto.colSet}
\title{Generate color set automatically}
\usage{
auto.colSet(n = 2, name = "Set1")
}
\arguments{
\item{n}{integer; number of colors needed}

\item{name}{character; Palette name}
}
\value{
an vector contains the color codes
}
\description{
Generate color set automatically
}
