% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{auto.point.size}
\alias{auto.point.size}
\title{Determine point size automatically}
\usage{
auto.point.size(n)
}
\arguments{
\item{n}{number of points to plot}
}
\value{
points' cex
}
\description{
Determine point size automatically
}
