% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{binarizeExp}
\alias{binarizeExp}
\title{binarize the expression value using the distribution}
\usage{
binarizeExp(x, out.prefix = NULL, G = NULL, e.TH = NULL,
  e.name = "Exp", verbose = F, draw.CI = T, zero.as.low = T, ...)
}
\arguments{
\item{x}{object; vector}

\item{out.prefix}{character; output prefix [default: NULL]}

\item{G}{integer; number of components [default: NULL]}

\item{e.TH}{double; for plot purpose: comparing the components with the user defined threhold [default: NULL]}

\item{e.name}{character; name of the expression metric [default: "Exp"]}

\item{verbose}{logical; [default: F]}

\item{draw.CI}{logical; [default: T]}

\item{zero.as.low}{logical; [default: T]}

\item{...}{parameters passed to plot.densityMclust}
}
\description{
binarize the expression value using the distribution
}
\details{
use mixture model to classify each data point. If G is null, the largest component is corresponding to binarized expression 1, and other components are corresponding to binarized expressed 0.
}
