% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{classify.outlier}
\alias{classify.outlier}
\title{outlier detection using extremevalues}
\usage{
classify.outlier(x, out.prefix = NULL)
}
\arguments{
\item{x}{object; vector}

\item{out.prefix}{character; output prefix [default: NULL]}
}
\description{
outlier detection using extremevalues
}
\details{
outlier detection using extremevalues
}
