% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{classifyCell.by.model}
\alias{classifyCell.by.model}
\title{classify cells using pre-trained model}
\usage{
classifyCell.by.model(
  obj.list,
  mod.rf,
  assay.name = "exprs",
  out.prefix = NULL,
  adjB = NULL,
  meta.info.tb = NULL,
  pca.rotation = NULL,
  pad.missing = F,
  allZeroAsLow = F,
  verbose = F,
  ncores = 16,
  ...
)
}
\arguments{
\item{obj.list}{object; named list of object of \code{singleCellExperiment} class}

\item{mod.rf}{object; pre-trained model used for prediction}

\item{assay.name}{character; which assay (default: "exprs")}

\item{out.prefix}{character; output prefix (default: NULL).}

\item{adjB}{character; batch column of the colData(obj). (default: NULL)}

\item{meta.info.tb}{data.frame; cell information table (default: NULL)}

\item{pca.rotation}{matrix; rotation matrix (default: NULL)}

\item{pad.missing}{logical; (default: FALSE)}

\item{allZeroAsLow}{logical; (default: FALSE)}

\item{verbose}{logical; (default: FALSE)}

\item{ncores}{integer; number of CPU to used (default: 16)}

\item{...}{parameters passed to some method}
}
\description{
classify cells using pre-trained model
}
\details{
classify cells using pre-trained model, which can be obtained by run classifyCell.by.sigGene()
}
