% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{classifyCell.by.sigGene}
\alias{classifyCell.by.sigGene}
\title{plot genes expression in pairs of clusters to examine the correlation}
\usage{
classifyCell.by.sigGene(obj.list, gene.desc.top, assay.name = "exprs",
  out.prefix = NULL, adjB = NULL, meta.cluster = NULL,
  sig.prevelance = 0.5)
}
\arguments{
\item{obj.list}{object; named list of object of \code{singleCellExperiment} class}

\item{gene.desc.top}{data.frame; signature genes}

\item{assay.name}{character; which assay (default: "exprs")}

\item{out.prefix}{character; output prefix (default: NULL).}

\item{adjB}{character; batch column of the colData(obj). (default: NULL)}

\item{sig.prevelance}{double; (default: 0.5)}
}
\description{
plot genes expression in pairs of clusters to examine the correlation
}
\details{
classify cells using the signature genes
}
