% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{collapseEffectSizeLong}
\alias{collapseEffectSizeLong}
\title{collapse effect size, give input table and group variate}
\usage{
collapseEffectSizeLong(
  dat.long,
  mode.collapse = "comb",
  group.2nd = "group.2nd",
  th.adj.P = 0.01,
  th.dprime = 0.15
)
}
\arguments{
\item{dat.long}{data.table; these columns are required: "geneID","aid","P.Value","adj.P.Val","sig","dprime","vardprime","group.2nd"}

\item{mode.collapse}{character; collapse method. one of "min", "comb" [default: "comb"]}

\item{group.2nd}{character; column of obj, used for collapse [default: NULL]}

\item{th.adj.P}{double; threshold [default: 0.01]}

\item{th.dprime}{double; threshold [default: 0.15]}
}
\value{
a gene table
}
\description{
collapse effect size, give input table and group variate
}
\details{
collapse effect size
}
