% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{comb.pvalue}
\alias{comb.pvalue}
\title{combine p values}
\usage{
comb.pvalue(p.mtx, method = "Fisher")
}
\arguments{
\item{p.mtx}{matrix; rows for genes, columns for studies.}

\item{method}{character; which method to use (default: "Fisher")}
}
\description{
combine p values
}
\details{
combine p values. Available method: Fisher's method
}
