% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.metaMA.R
\name{directEScombi}
\alias{directEScombi}
\title{directional Effect Size combination}
\usage{
directEScombi(ES, varES, useREM = TRUE)
}
\arguments{
\item{ES}{data frame or matrix; row for genes, column for studies}

\item{varES}{data frame or matrix; row for genes, column for studies}

\item{useREM}{logical; . [default TRUE]}
}
\value{
a matrix
}
\description{
directional Effect Size combination
}
\details{
calculate the weight averaged effect size from multiple studies
}
