% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.metaMA.R
\name{directEScombiFromLongTable}
\alias{directEScombiFromLongTable}
\title{directional Effect Size combination}
\usage{
directEScombiFromLongTable(
  dat.long,
  idx.geneID = "geneID",
  idx.aid = "aid",
  idx.ES = "dprime",
  idx.varES = "vardprime",
  ...
)
}
\arguments{
\item{dat.long}{data.table; input data}

\item{idx.geneID}{character; column for geneID (default: "geneID")}

\item{idx.aid}{character; column for aid (default: "aid")}

\item{idx.ES}{character; column for ES (default: "dprime")}

\item{idx.varES}{character; column for varES (default: "vardprime")}

\item{...}{parameters passed to directEScombi()}
}
\value{
a data.table
}
\description{
directional Effect Size combination
}
\details{
wrapper function to run directEScombi()
}
