% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.metaMA.R
\name{effectsize}
\alias{effectsize}
\title{effect size from moderate t test}
\usage{
effectsize(tstat, ntilde, m)
}
\arguments{
\item{tstat}{vector; t value of limma's moderate t test}

\item{ntilde}{doule; .}

\item{m}{doule; .}
}
\value{
a matrix
}
\description{
effect size from moderate t test
}
\details{
calculate the effect size
}
