% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expressedFraction}
\alias{expressedFraction}
\title{For each gene, calculate the frequency of cells in each clusters are expressed.}
\usage{
expressedFraction(exp.bin, group, n.cores = NULL)
}
\arguments{
\item{exp.bin}{numeric; binarized expression matrix, rows for genes and columns for samples. value 1 means expressed.}

\item{group}{character; clusters of the samples belong to}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}
}
\description{
For each gene, calculate the frequency of cells in each clusters are expressed.
}
