% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findDEGenesByAOV}
\alias{findDEGenesByAOV}
\title{differential expression analysis}
\usage{
findDEGenesByAOV(
  xdata,
  xlabel,
  batch = NULL,
  out.prefix = NULL,
  pmod = NULL,
  F.FDR.THRESHOLD = 0.01,
  HSD.FDR.THRESHOLD = 0.01,
  HSD.FC.THRESHOLD = 1,
  use.Kruskal = F,
  verbose = F,
  n.cores = NULL,
  gid.mapping = NULL
)
}
\arguments{
\item{xdata}{data frame or matrix; rows for genes and columns for samples}

\item{xlabel}{factor; cluster label of the samples, with length equal to the number of columns in xdata}

\item{batch}{factor; covariate. (default: NULL)}

\item{out.prefix}{character; if not NULL, write the result to the file(s). (default: NULL)}

\item{pmod}{character;}

\item{F.FDR.THRESHOLD}{numeric; threshold of the adjusted p value of F-test. (default: 0.01)}

\item{HSD.FDR.THRESHOLD}{numeric; threshold of the adjusted p value of HSD-test (default: 0.01)}

\item{HSD.FC.THRESHOLD}{numeric; threshold of the absoute diff of HSD-test (default: 1)}

\item{use.Kruskal}{logical; whether use Kruskal test for ranking genes (default: FALSE)}

\item{verbose}{logical; whether output all genes' result. (default: F)}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{gid.mapping}{named character; gene id to gene symbol mapping. (default: NULL)}
}
\value{
List with the following elements:
\item{aov.out}{data.frame, test result of all genes (rownames of xdata)}
\item{aov.out.sig}{format as aov.out, but only significant genes. }
}
\description{
differential expression analysis
}
