% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findKneePoint}
\alias{findKneePoint}
\title{Find the knee point of the scree plot}
\usage{
findKneePoint(pcs)
}
\arguments{
\item{pcs}{principal component values sorted decreasingly}
}
\value{
index of the knee plot
}
\description{
Find the knee point of the scree plot
}
\details{
Given sorted decreasingly PCs, find the knee point which have the largest distance to
the line defined by the first point and the last point in the scree plot
}
