% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAUC}
\alias{getAUC}
\title{calculate the AUC of one gene, using it as a classifier. code from SC3}
\usage{
getAUC(gene, labels, use.rank = T, geneID = "GeneX", method.Max = "mean")
}
\arguments{
\item{gene}{numeric; expression profile of one gene across samples}

\item{labels}{character; clusters of the samples belong to}

\item{use.rank}{logical; using the expression value itself or convert to rank value. (default: TRUE)}

\item{geneID}{character; geneID}

\item{method.Max}{character; method to find highest group, one of "mean", "median", "rank.mean" (default: mean)}
}
\description{
calculate the AUC of one gene, using it as a classifier. code from SC3
}
