% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate.by.avg}
\alias{integrate.by.avg}
\title{integrate dataset by calculating the average expression and then clustering the pseudo bulk data}
\usage{
integrate.by.avg(sce.list, out.prefix, assay.name = "exprs",
  is.avg = FALSE, ncores = 6, use.deg = TRUE, gene.de.list = NULL,
  avg.by = "majorCluster", n.downsample = NULL, topGene.lo = -1.5,
  topGene.hi = 1.5, method.avg = "zscore", ...)
}
\arguments{
\item{sce.list}{list; list of object of \code{singleCellExperiment} class}

\item{out.prefix}{character; output prefix.}

\item{assay.name}{character; which assay (default: "exprs")}

\item{is.avg}{logical; whether the objects in sce.list are average expression (default: FALSE)}

\item{ncores}{integer; number of cores to use (default: 6)}

\item{use.deg}{logical; whether use only the differentially expressed genes (default: FALSE)}

\item{gene.de.list}{list; if not NULL, each element is a data.frame with "geneID" column (default: NULL)}

\item{avg.by}{character; calculate the average expression of cells group by the specifid column (default: "majorCluster")}

\item{n.downsample}{integer; number of cells in each cluster to downsample to (default: NULL)}

\item{topGene.lo}{double; for top gene heatmap.(default: -1.5)}

\item{topGene.hi}{double; for top gene heatmap.(default: 1.5)}

\item{method.avg}{character; method of calculate the average expression. Passed to `avg` of `ssc.average.cell`.(default: "zscore")}

\item{...}{parameters passed to ssc.clusterMarkerGene}
}
\description{
integrate dataset by calculating the average expression and then clustering the pseudo bulk data
}
\details{
method to calculate the average expression can be one of "mean", "zscore"
}
