% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{matrix.waterfall}
\alias{matrix.waterfall}
\title{make the matrix looks like "waterfall" (typically genes expression)}
\usage{
matrix.waterfall(
  dat,
  score.alpha = 1.5,
  clust.row = FALSE,
  clust.column = FALSE,
  waterfall.column = F,
  waterfall.row = F,
  type.return = "matrix",
  do.norm = T
)
}
\arguments{
\item{dat}{matrix; matrix}

\item{score.alpha}{double; for row/column score; (default: 1.5)}

\item{clust.row}{logical, character or dendrogram; passed to cluster_rows of Heatmap (default: FALSE)}

\item{clust.column}{logical, character or dendrogram; passed to cluster_columns of Heatmap (default: FALSE)}

\item{waterfall.column}{logical, order rows to make plot like waterfall (default: FALSE)}

\item{waterfall.row}{logical, order rows to make plot like waterfall (default: FALSE)}

\item{type.return}{character; return type; (default: "matrix")}

\item{do.norm}{logical; normalized the data before calculating scores?; (default: TRUE)}
}
\description{
make the matrix looks like "waterfall" (typically genes expression)
}
