% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcrf}
\alias{mcrf}
\title{Wraper for running random forest classifier (multiple core version)}
\usage{
mcrf(ntree, ncore, ...)
}
\arguments{
\item{ntree}{integer; Number of trees to grow.}

\item{ncore}{integer; Number of cores to use.}

\item{...}{passed to randomForest::randomForest}
}
\description{
Wraper for running random forest classifier (multiple core version)
}
