% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.branch}
\alias{plot.branch}
\title{plot matrix (typically genes expression)}
\usage{
\method{plot}{branch}(obj.clust, out.prefix, ncls = 1, cluster = NULL)
}
\arguments{
\item{obj.clust}{object;}

\item{out.prefix}{character; output prefix.}

\item{ncls}{integer; (default: 1)}

\item{cluster}{integer vector; (default: NULL)}
}
\description{
plot matrix (typically genes expression)
}
\details{
plot dendrogram
}
