% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.matrix.simple}
\alias{plot.matrix.simple}
\title{plot matrix (typically genes expression)}
\usage{
\method{plot}{matrix.simple}(dat, out.prefix = NULL, mytitle = "",
  show.number = TRUE, do.clust = FALSE, z.lo = NULL, z.hi = NULL,
  palatte = NULL, pdf.width = 8, pdf.height = 8,
  exp.name = "Count")
}
\arguments{
\item{dat}{matrix; matrix}

\item{out.prefix}{character; output prefix.}

\item{mytitle}{character; (default: "")}

\item{show.number}{logical; (default: TRUE)}

\item{do.clust}{logical or dendrogram; passed to cluster_columns and cluster_rows of Heatmap (default: FALSE)}

\item{z.lo}{double; (default: NULL)}

\item{z.hi}{double; (default: NULL)}

\item{palatte}{character; (default: NULL)}

\item{pdf.width}{double; width of the output plot (default: 22)}

\item{pdf.height}{double; height of the output plot (default: 22)}

\item{exp.name}{character; showd in the legend (default: "Count")}
}
\description{
plot matrix (typically genes expression)
}
\details{
plot matrix
}
