% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{plot.pairs.cor}
\alias{plot.pairs.cor}
\title{plot genes expression in pairs of clusters to examine the correlation}
\usage{
\method{plot}{pairs.cor}(
  sce.pb,
  assay.name,
  x.cluster,
  out.prefix,
  genes = NULL,
  gene.highlight = NULL,
  plot.ncol = 7,
  plot.width = 22,
  plot.height = 22
)
}
\arguments{
\item{sce.pb}{object; object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{x.cluster}{character; cluster to be showed in x axis}

\item{out.prefix}{character; output prefix.}

\item{genes}{vector; genes to used to calculate the correlation and make the scatter plot (default: NULL)}

\item{gene.highlight}{vector; genes to highlight (default: NULL)}

\item{plot.ncol}{integer; number of columns in facet_wrap() (default: 7)}

\item{plot.width}{double; width of the output plot (default: 22)}

\item{plot.height}{double; height of the output plot (default: 22)}
}
\description{
plot genes expression in pairs of clusters to examine the correlation
}
\details{
used to examine the correlation between pairs of clusters
}
