% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{plotSigGene}
\alias{plotSigGene}
\title{plot signature genes}
\usage{
plotSigGene(
  obj.list,
  gene.core.tb,
  out.prefix,
  assay.name = "exprs",
  adjB = NULL,
  meta.cluster = NULL,
  meta.info.tb = NULL,
  meta.col = "metaCluster",
  p.ncol = 2,
  pdf.width = NULL,
  pdf.height = NULL,
  verbose = F,
  ptype = "heatmap",
  ncores = 16,
  ...
)
}
\arguments{
\item{obj.list}{object; named list of object of \code{singleCellExperiment} class}

\item{gene.core.tb}{data.frame; signature genes}

\item{out.prefix}{character; output prefix}

\item{assay.name}{character; which assay (default: "exprs")}

\item{adjB}{character; batch column of the colData(obj). (default: NULL)}

\item{meta.cluster}{character; (default: NULL)}

\item{meta.info.tb}{data.frame; cell information table (default: NULL)}

\item{meta.col}{character; column in meta.info.tb, by which to group cells (default: "meta.cluster")}

\item{p.ncol}{integer; number of columns in the violin plot (default: 2)}

\item{pdf.width}{double; width of the output plot (default: NULL)}

\item{pdf.height}{double; height of the output plot (default: NULL)}

\item{verbose}{logical; (default: FALSE)}

\item{ptype}{character; (default: "heatmap")}

\item{ncores}{integer; number of CPU to used (default: 16)}

\item{...}{parameters passed to plotting methods}
}
\description{
plot signature genes
}
\details{
make some plot(s) of signature genes
}
