% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{rank.de.gene}
\alias{rank.de.gene}
\title{get gene ranking table from obj}
\usage{
rank.de.gene(
  obj,
  group = "pca.SNN.kauto",
  sort.by = "median.rank",
  weight.adj = NULL
)
}
\arguments{
\item{obj}{object; object of \code{singleCellExperiment} class}

\item{group}{character; columan name in colData(obj)}

\item{sort.by}{character; by which to sort the genes [default: "median.rank"]}
}
\value{
a gene table
}
\description{
get gene ranking table from obj
}
\details{
rank genes
}
