% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ssc.plot.silhouette}
\alias{ssc.plot.tsne}
\alias{ssc.plot.violin}
\alias{ssc.plot.pca}
\alias{ssc.plot.cor}
\alias{ssc.plot.heatmap}
\alias{ssc.plot.GeneDensity}
\alias{ssc.assay.hclust}
\alias{ssc.assay.zscore}
\alias{ssc.order}
\alias{ssc.average.cell}
\alias{ssc.displayName2id}
\alias{ssc.id2displayName}
\alias{ssc.downsample}
\alias{ssc.toLongTable}
\alias{ssc.moduleScore}
\alias{ssc.scale}
\alias{ssc.build}
\alias{effectsize}
\alias{directEScombi}
\alias{directEScombiFromLongTable}
\alias{collapseEffectSizeLong}
\alias{simple.removeBatchEffect}
\alias{run.cutree}
\alias{run.cutreeDynamic}
\alias{loginfo}
\alias{cor.BLAS}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sscVis}{\code{\link[sscVis]{collapseEffectSizeLong}}, \code{\link[sscVis]{cor.BLAS}}, \code{\link[sscVis]{directEScombi}}, \code{\link[sscVis]{directEScombiFromLongTable}}, \code{\link[sscVis]{effectsize}}, \code{\link[sscVis]{loginfo}}, \code{\link[sscVis]{run.cutree}}, \code{\link[sscVis]{run.cutreeDynamic}}, \code{\link[sscVis]{simple.removeBatchEffect}}, \code{\link[sscVis]{simple.removeBatchEffect}}, \code{\link[sscVis]{ssc.assay.hclust}}, \code{\link[sscVis]{ssc.assay.zscore}}, \code{\link[sscVis]{ssc.average.cell}}, \code{\link[sscVis]{ssc.build}}, \code{\link[sscVis]{ssc.displayName2id}}, \code{\link[sscVis]{ssc.downsample}}, \code{\link[sscVis]{ssc.id2displayName}}, \code{\link[sscVis]{ssc.moduleScore}}, \code{\link[sscVis]{ssc.order}}, \code{\link[sscVis]{ssc.plot.cor}}, \code{\link[sscVis]{ssc.plot.GeneDensity}}, \code{\link[sscVis]{ssc.plot.heatmap}}, \code{\link[sscVis]{ssc.plot.pca}}, \code{\link[sscVis]{ssc.plot.silhouette}}, \code{\link[sscVis]{ssc.plot.tsne}}, \code{\link[sscVis]{ssc.plot.violin}}, \code{\link[sscVis]{ssc.scale}}, \code{\link[sscVis]{ssc.toLongTable}}}
}}

