% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.cutree}
\alias{run.cutree}
\title{run cutree on rows of the given matrix}
\usage{
run.cutree(
  dat,
  method.hclust = "ward.D2",
  method.distance = "spearman",
  k = 1,
  ...
)
}
\arguments{
\item{method.hclust}{character; clustering method for hclust [default: "ward.D2"]}

\item{method.distance}{character; distance method for hclust [default: "spearman"]}

\item{...}{parameters passed to cutree}

\item{data}{data frame or matrix;}
}
\value{
a matrix with dimention as input ( samples in rows and variables in columns)
}
\description{
run cutree on rows of the given matrix
}
\details{
cutree on rows of the given matrix
}
