% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.cutreeDynamic}
\alias{run.cutreeDynamic}
\title{run dynamicTreeCut::cutreeDynamic on rows of the given matrix}
\usage{
run.cutreeDynamic(
  dat,
  method.hclust = "ward.D2",
  deepSplit = 4,
  minClusterSize = 2,
  ...
)
}
\arguments{
\item{deepSplit}{integer; passed to dynamicTreeCut::cutreeDynamic}

\item{minClusterSize}{integer; passed to dynamicTreeCut::cutreeDynamic}

\item{...}{parameters passed to dynamicTreeCut::cutreeDynamic}

\item{data}{data frame or matrix;}
}
\value{
a matrix with dimention as input ( samples in rows and variables in columns)
}
\description{
run dynamicTreeCut::cutreeDynamic on rows of the given matrix
}
\details{
dynamicTreeCut::cutreeDynamic on rows of the given matrix
}
