% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.DE.matrix}
\alias{run.DE.matrix}
\title{run DE given an expression matrix}
\usage{
run.DE.matrix(
  xdata,
  xlabel,
  batch = NULL,
  out.prefix = NULL,
  ncell.downsample = NULL,
  T.fdr = 0.05,
  T.logFC = 1,
  n.cores = NULL,
  group = NULL,
  gid.mapping = NULL,
  rn.seed = 9999,
  method = "lm"
)
}
\arguments{
\item{xdata}{data frame or matrix; rows for genes and columns for samples}

\item{xlabel}{factor; cluster label of the samples, with length equal to the number of columns in xdata}

\item{batch}{factor; covariate. (default: NULL)}

\item{out.prefix}{character; if not NULL, write the result to the file(s). (default: NULL)}

\item{ncell.downsample}{integer; for each group, number of cells downsample to. (default: NULL)}

\item{T.fdr}{numeric; threshold of the adjusted p value of moderated t-test (default: 0.05)}

\item{T.logFC}{numeric; threshold of the absoute diff (default: 1)}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{group}{character; group of interest, if NULL the last group will be used (default: NULL)}

\item{gid.mapping}{named character; gene id to gene symbol mapping. (default: NULL)}

\item{rn.seed}{integer; random number seed (default: 9999)}

\item{method}{character; . (default: "lm")}
}
\value{
a matrix with dimention as input ( samples in rows and variables in columns)
}
\description{
run DE given an expression matrix
}
\details{
diffeerentially expressed genes dectection using naive methods
}
