% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.KNN}
\alias{run.KNN}
\title{Wraper for running random forest classifier}
\usage{
run.KNN(xdata, xlabel, ydata, k = 1)
}
\arguments{
\item{xdata}{data frame or matrix; data used for training, with sample id in rows and variables in columns}

\item{xlabel}{factor; classification label of the samples, with length equal to the number of rows in xdata}

\item{ydata}{data frame or matrix; data to be predicted the label, same format as xdata}

\item{k}{parameter k of function knn() (default: 1)}
}
\value{
List with the following elements:
\item{ylabel}{ppredicted labels of the samples in ydata}
}
\description{
Wraper for running random forest classifier
}
