% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.SVM}
\alias{run.SVM}
\title{Wraper for running svm}
\usage{
run.SVM(xdata, xlabel, ydata, kern = "linear")
}
\arguments{
\item{xdata}{data frame or matrix; data used for training, with sample id in rows and variables in columns}

\item{xlabel}{factor; classification label of the samples, with length equal to the number of rows in xdata}

\item{ydata}{data frame or matrix; data to be predicted the label, same format as xdata}

\item{kern}{character; which kernel to use, can be one of linear, polynomial, radial and sigmoid (default: "linear")}
}
\value{
List with the following elements:
\item{ylabel}{ppredicted labels of the samples in ydata}
\item{rfsel}{trained model; output of varSelRF()}
}
\description{
Wraper for running svm
}
