% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\docType{data}
\name{sce.Pollen}
\alias{sce.Pollen}
\title{Single cell RNA-Seq data extracted from a publication by Pollen et al.}
\source{
\url{https://www.nature.com/articles/nbt.2967}

Expression data contained in a \code{SingleCellExperiment} object
}
\description{
Single cell RNA-Seq data extracted from a publication by Pollen et al.
}
