% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple.removeBatchEffect}
\alias{simple.removeBatchEffect}
\title{Modified from limma::removeBatchEffect, a little different design matrix}
\usage{
simple.removeBatchEffect(x, batch = NULL, covariates = NULL, ...)
}
\arguments{
\item{x}{matrix; samples in columns and variables in rows}

\item{batch}{character; batch vector (default: NULL)}

\item{covariates}{double; other covariates to adjust (default: NULL)}

\item{...}{parameters passed to lmFit}
}
\value{
a matrix with dimention as input ( samples in rows and variables in columns)
}
\description{
Modified from limma::removeBatchEffect, a little different design matrix
}
\details{
Modified from limma::removeBatchEffect, a little different design matrix
}
