% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.assay.hclust}
\alias{ssc.assay.hclust}
\title{sort by hierarchical clustering}
\usage{
ssc.assay.hclust(obj, assay.name = "exprs", order.col = FALSE,
  order.row = FALSE, clustering.distance = "spearman",
  clustering.method = "complete", k.row = 1, k.col = 1)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{order.col}{logical; wheter order columns (default: FALSE)}

\item{order.row}{logical; wheter order row (default: FALSE)}

\item{clustering.distance}{character; one of spearmn, pearson and euclidean (default: "spearman")}

\item{clustering.method}{character; method for hclust (default: "complete")}

\item{k.row}{integer; number of clusters in the rows (default: 1)}

\item{k.col}{integer; number of clusters in the columns (default: 1)}
}
\description{
sort by hierarchical clustering
}
\details{
order genes or cells according the assay, using hclust.
}
