% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.assay.zscore}
\alias{ssc.assay.zscore}
\title{scale the assay per gene}
\usage{
ssc.assay.zscore(
  obj,
  assay.name = "exprs",
  assay.new = NULL,
  covar = "patient",
  n.cores = NULL,
  z.lo = NULL,
  z.hi = NULL
)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{assay.new}{character; assay name to store the scaled- expression;if NULL, will be (assay.name).z (default: NULL)}

\item{covar}{character; perform scale in cells grouping by covar (default: "patient")}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{z.lo}{double; z-score lower boundary; if set, z-score lower than this will be set to this (default: NULL)}

\item{z.hi}{double; z-score higher boundary; if set, z-score higher than this will be set to this (default: NULL)}
}
\description{
scale the assay per gene
}
