% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.average.cell}
\alias{ssc.average.cell}
\title{calculate the average expression of cells}
\usage{
ssc.average.cell(obj, assay.name = "exprs", columns = "majorCluster")
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{columns}{character; columns of colData(obj) used for grouping (default: "majorCluster")}
}
\description{
calculate the average expression of cells
}
