% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.build}
\alias{ssc.build}
\title{Build an SingleCellExperiment object}
\usage{
ssc.build(x, display.name = NULL)
}
\arguments{
\item{x}{matrix/data.frame or SingleCellExperiment; input expression data}

\item{display.name}{a vector, should be human readable gene name}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
Build an SingleCellExperiment object from a matrix or data frame
}
\details{
if x is an object of SingleCellExperiment, just clear the metadata;
if x is matrix/data.frame, convert it to an object of SingleCellExperiment.
Also a vector `display.name` can be provided, which would be used in some plots,
such as geneOnTSNE, heatmap. The row names of SingleCellExperiment object usually be gene id
(e.g. entrez ID, Ensemble ID), the `display.name` should be human readable gene name (
e.g. HGNC gene symbol). If `display.name` is NULL (default), the row names of SingleCellExperiment object
would be used.
}
