% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{ssc.clust}
\alias{ssc.clust}
\title{Perform clustering using reduced data}
\usage{
ssc.clust(obj, assay.name = "exprs", method.reduction = "iCor",
  method = "kmeans", k.batch = 2:6, method.vgene = "sd")
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{method.reduction}{character; which dimention reduction method to be used, should be one of
"iCor", "pca" and "none". (default: "iCor")}

\item{method}{character; clustering method to be used, should be one of "kmeans" and "hclust". (default: "kmeans")}

\item{k.batch}{integer; number of clusters to be evaluated. (default: 2:6)}

\item{method.vgene}{character; variable gene identification method used. (default: "sd")}
}
\value{
an object of \code{SingleCellExperiment} class with cluster labels added.
}
\description{
Perform clustering using reduced data
}
\details{
If no dimension reduction performed or method is "none", expression data of variable genes,
which can be speficed by method.vgene, will be used for clustering. Otherwise, the reduced data specified by
method.reduction will be used. The cluster label will stored in the colData of the object
of \code{singleCellExperiment} class, with colname in the format of \{method.reduction\}.\{method\}k\{k\}
where \{k\} get value(s) from k.batch.
}
