% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.clusterMarkerGene}
\alias{ssc.clusterMarkerGene}
\title{identify marker genes of each cluster}
\usage{
ssc.clusterMarkerGene(
  obj,
  assay.name = "exprs",
  ncell.downsample = NULL,
  group.var = "majorCluster",
  batch = NULL,
  assay.bin = NULL,
  out.prefix = NULL,
  n.cores = NULL,
  do.plot = T,
  F.FDR.THRESHOLD = 0.01,
  pairwise.P.THRESHOLD = 0.01,
  pairwise.FC.THRESHOLD = 1,
  use.Kruskal = F,
  use.medianMax = F,
  do.force = F,
  verbose = F,
  ...
)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{ncell.downsample}{integer; for each group, number of cells downsample to. (default: NULL)}

\item{group.var}{character; column in the colData(obj) used for grouping. (default: "majorCluster")}

\item{batch}{character; covariate. (default: NULL)}

\item{assay.bin}{character; binarized expression assay (default: NULL)}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{do.plot}{logical; whether plot. (default: TRUE)}

\item{F.FDR.THRESHOLD}{numeric; threshold of the adjusted p value of F-test. (default: 0.01)}

\item{pairwise.P.THRESHOLD}{numeric; threshold of the adjusted p value of HSD-test (default: 0.01)}

\item{pairwise.FC.THRESHOLD}{numeric; threshold of the absoute diff of HSD-test (default: 1)}

\item{use.Kruskal}{logical; whether use Kruskal test for ranking genes (default: FALSE)}

\item{use.medianMax}{logical; whether use median for gene classification (default: FALSE)}

\item{do.force}{logical; . (default: FALSE)}

\item{verbose}{logical; whether output all genes' result. (default: FALSE)}

\item{...}{parameters passed to findDEGenesByAOV}
}
\description{
identify marker genes of each cluster
}
\details{
identify marker genes based on aov and AUC.
}
