% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.clustSubsamplingClassification}
\alias{ssc.clustSubsamplingClassification}
\title{Clustering with subsampling and classification}
\usage{
ssc.clustSubsamplingClassification(obj, assay.name = "exprs", frac = 0.4,
  method.vgene = "sd", method.reduction = "iCor", method.clust = "kmeans",
  method.classify = "knn", pca.npc = NULL, iCor.niter = 1, use.proj = T,
  vis.proj = F, ncore = NULL, k.batch = 2:6, seed = NULL)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{frac}{numeric; subsample to frac of original samples. (default: 0.4)}

\item{method.vgene}{character; variable gene identification method used. (default: "sd")}

\item{method.reduction}{character; which dimention reduction method to be used, should be one of
"iCor", "pca" and "none". (default: "iCor")}

\item{method.clust}{character; clustering method to be used, should be one of "kmeans" and "hclust". (default: "kmeans")}

\item{method.classify}{character; method used for classification, one of "knn" and "RF". (default: "knn")}

\item{pca.npc}{integer; number of pc be used. Only for reduction method "pca". (default: NULL)}

\item{iCor.niter}{integer; number of iteration of calculating the correlation. Used in reduction method "iCor". (default: 1)}

\item{use.proj}{logical; whether use the projected data for classification. (default: T)}

\item{vis.proj}{logical; whether get low dimensional representation for visualization. (default: F)}

\item{ncore}{integer; number of cpu to use. (default: NULL)}

\item{k.batch}{integer; number of clusters to be evaluated. (default: 2:6)}

\item{seed}{integer; seed of random number generation. (default: NULL)}
}
\value{
an object of \code{SingleCellExperiment} class with cluster labels added.
}
\description{
Clustering with subsampling and classification
}
\details{
The function first subsmaple the samples to the specified fraction (such 40%), and perform clustering. The clustering will
make labels for the subsampled samples. Using the labels, original data or projected data via the method specified in
"method.reduction" will be used for trainning a classifier. Then the classifier will predict the labels of the samples not subsampled,
using original data or projected data dependent on the option use.proj. The final cluster labels combining
that of bath sampled and unsampled samples, will stored in the colData of the object of \code{singleCellExperiment} class,
with colname in the format of \{method.reduction\}.\{method\}k\{k\} where \{k\} get value(s) from k.batch.
}
