% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.DEGene.limma}
\alias{ssc.DEGene.limma}
\title{identify differential genes of each cluster (comparing the cluster with all others), using limma}
\usage{
ssc.DEGene.limma(obj, assay.name = "exprs", ncell.downsample = NULL,
  group.var = "majorCluster", group.list = NULL, batch = NULL,
  out.prefix = NULL, n.cores = NULL, do.plot = T, T.fdr = 0.01,
  T.logFC = 1, verbose = F)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{ncell.downsample}{integer; for each group, number of cells downsample to. (default: NULL)}

\item{group.var}{character; column in the colData(obj) used for grouping. (default: "majorCluster")}

\item{group.list}{character; DEG of groups to calculate. If NULL, all groups. (default: "NULL")}

\item{batch}{character; covariate. (default: NULL)}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{do.plot}{logical; whether plot. (default: TRUE)}

\item{T.fdr}{numeric; threshold of the adjusted p value of moderated t-test (default: 0.05)}

\item{T.logFC}{numeric; threshold of the absoute diff (default: 1)}

\item{verbose}{logical; whether output all genes' result. (default: FALSE)}
}
\description{
identify differential genes of each cluster (comparing the cluster with all others), using limma
}
\details{
identify differential genes using limma
}
