% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{ssc.displayName2id}
\alias{ssc.displayName2id}
\title{Convert display name (gene symbol) to gene id}
\usage{
ssc.displayName2id(obj, display.name)
}
\arguments{
\item{obj}{object of \code{SingleCellExperiment} class}

\item{display.name}{character; disaply name}
}
\value{
If successfull vector contains the gene id; otherwise NULL
}
\description{
Convert display name (gene symbol) to gene id
}
