% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.downsample}
\alias{ssc.downsample}
\title{downsample}
\usage{
ssc.downsample(
  obj,
  ncell.downsample = NULL,
  group.var = "majorCluster",
  rn.seed = 9999
)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{ncell.downsample}{integer; for each group, number of cells downsample to. (default: NULL)}

\item{group.var}{character; column in the colData(obj) used for grouping. (default: "majorCluster")}

\item{rn.seed}{integer; random number seed (default: 9999)}
}
\description{
downsample
}
\details{
return a downsampled object of \code{singleCellExperiment} class.
}
