% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{ssc.id2displayName}
\alias{ssc.id2displayName}
\title{Convert gene id to display name (gene symbol)}
\usage{
ssc.id2displayName(obj, ids)
}
\arguments{
\item{obj}{object of \code{SingleCellExperiment} class}

\item{ids}{character; gene ids}
}
\value{
If successfull vector contains the disaply name; otherwise NULL
}
\description{
Convert gene id to display name (gene symbol)
}
