% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.moduleScore}
\alias{ssc.moduleScore}
\title{Add module scores for gene expression programs in single cells. Mofidy from Seurat::AddModuleScore}
\usage{
ssc.moduleScore(obj, features, pool = NULL, nbin = 24, ctrl = 100,
  assay.name = "exprs", seed = 1)
}
\arguments{
\item{obj}{object of SingleCellExperiment}

\item{features}{Feature expression programs in named list}

\item{pool}{List of features to check expression levels agains, defaults to \code{rownames(x = object)}}

\item{nbin}{Number of bins of aggregate expression levels for all analyzed features}

\item{ctrl}{Number of control features selected from the same bin per analyzed feature}

\item{assay.name}{Name of assay to use}

\item{seed}{Set a random seed}
}
\value{
Returns a SingleCellExperiment object with module scores added to object meta data
}
\description{
Calculate the average expression levels of each program (cluster) on single cell level,
subtracted by the aggregated expression of control feature sets.
All analyzed features are binned based on averaged expression, and the control features are
randomly selected from each bin.
}
\references{
Tirosh et al, Science (2016); Seurat's source code
}
