% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.order}
\alias{ssc.order}
\title{order genes and cells}
\usage{
ssc.order(obj, columns.order = NULL, gene.desc = NULL)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{columns.order}{character; columns of colData(obj) used for ordering (default: NULL)}

\item{gene.desc}{data.frame; it must contain columns geneID and Group (default: NULL)}
}
\description{
order genes and cells
}
