% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.plot.cor}
\alias{ssc.plot.cor}
\title{Plot correlation.}
\usage{
ssc.plot.cor(
  obj,
  feat1,
  feat2,
  type1 = "gene",
  type2 = "gene",
  assay.name = "exprs",
  legend.w = 1,
  legend.ncol = NULL,
  add.legend = F,
  out.prefix = NULL,
  vector.friendly = F,
  ...
)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{feat1}{character; feature on x axis}

\item{feat2}{character; feature on y axis}

\item{type1}{character; feature type on x axis. (default: "gene")}

\item{type2}{character; feature type on y axis. (default: "gene")}

\item{assay.name}{character; which assay (default: "exprs")}

\item{legend.w}{numeric; adjust legend width (default: 1)}

\item{legend.ncol}{integer;  (default: NULL)}

\item{add.legend}{logical;  (default: F)}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{vector.friendly}{logical; output vector friendly figure (default: FALSE)}

\item{...}{parameter passed to ggscatter}
}
\description{
Plot correlation.
}
