% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.plot.heatmap}
\alias{ssc.plot.heatmap}
\title{plot heatmap}
\usage{
ssc.plot.heatmap(obj, assay.name = "exprs", out.prefix = NULL,
  ncell.downsample = NULL, ave.by = NULL, columns = NULL,
  columns.order = NULL, gene.desc = NULL, colSet = list(),
  pdf.width = 16, pdf.height = 15, do.scale = TRUE, z.lo = -2.5,
  z.hi = 2.5, z.step = 1, exp.title = "Exp", do.clustering.row = T,
  do.clustering.col = T, dend.col = FALSE, dend.row = FALSE,
  clustering.distance = "spearman", clustering.method = "complete",
  k.row = 1, k.col = 1, palette.name = NULL, row.split = NULL,
  annotation_legend_param = list(), ann.bar.height = 1.5,
  mytitle = "", ...)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{ncell.downsample}{integer; number of cells downsample to. (default: NULL)}

\item{ave.by}{character; average the expression profile grouping by this. (default: NULL)}

\item{columns}{character; columns in colData(obj) to be showd. must be subset of columns of colData(obj) and ave.by (if it's not NULL) (default: NULL)}

\item{columns.order}{character; columns of colData(obj) used for ordering (default: NULL)}

\item{gene.desc}{data.frame; it must contain columns geneID and Group (default: NULL)}

\item{colSet}{list; mapping iterms in the names to colors in the values. (default: list())}

\item{pdf.width}{double; width of the pdf file. (default:16)}

\item{pdf.height}{double; height of the pdf file. (default:15)}

\item{do.scale}{logical; wheter scale the rows, just for visualization. (default: TRUE)}

\item{z.lo}{double; z-score lower boundary; z-score lower than this will be set to this (default: -2.5)}

\item{z.hi}{double; z-score higher boundary; z-score higher than this will be set to this (default: 2.5)}

\item{z.step}{double; z-score step, used for coloring the expression value (default: 1)}

\item{exp.title}{character; title for the expression legend (default: "Exp")}

\item{do.clustering.row}{logical; wheter order row (default: TRUE)}

\item{do.clustering.col}{logical; wheter order columns (default: TRUE)}

\item{dend.col}{dendrogram of the columns, 'cluster_columns' of ComplexHeatmap::Heatmap  (default: FALSE)}

\item{dend.row}{dendrogram of the rows, 'cluster_rows' of ComplexHeatmap::Heatmap (default: FALSE)}

\item{clustering.distance}{character; one of spearmn, pearson and euclidean (default: "spearman")}

\item{clustering.method}{character; method for hclust (default: "complete")}

\item{k.row}{integer; number of clusters in the rows (default: 1)}

\item{k.col}{integer; number of clusters in the columns (default: 1)}

\item{palette.name}{character; which palette to use, such as "RdBu","RdYlBu" (default: NULL)}

\item{row.split}{vector; used for row; (default: NULL)}

\item{annotation_legend_param}{list; (default: list())}

\item{ann.bar.height}{double; height of the top annotation (default: 1.5)}

\item{mytitle}{character; title of the figure (default: "")}

\item{...}{parameters pass to sscClust:::plot.matrix.simple()}
}
\description{
plot heatmap
}
\details{
identify marker genes based on aov and AUC.
}
