% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.plot.pca}
\alias{ssc.plot.pca}
\title{Plot pca result, such as scree plot.}
\usage{
ssc.plot.pca(obj, out.prefix = NULL, p.ncol = 2)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{p.ncol}{integer; number of columns in the figure layout. (default: 2)}
}
\description{
Plot pca result, such as scree plot.
}
