% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.plotGeneDensity}
\alias{ssc.plotGeneDensity}
\title{plot gene expression density}
\usage{
ssc.plotGeneDensity(obj, out.prefix, gene.id, gene.symbol,
  assay.name = "norm_exprs", expT = c(0.3, 0.3),
  pallete.name = "heat", ann.txt.dis = 0.3, ann.txt.cex = 1.2,
  my.title = "")
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{out.prefix}{character; output prefix. required}

\item{gene.id}{should be in rownames(obj); genes to plot}

\item{gene.symbol}{should be in rowData(obj)[,"display.name"]; genes to plot}

\item{assay.name}{character; which assay (default: "exprs")}

\item{expT}{double; expression threshold of the genes. (default: c(0.3,0.3))}

\item{pallete.name}{character; pallete to use. (default: "heat")}

\item{ann.txt.dis}{double; adjust the position of the annotation text (default: 0.3)}

\item{ann.txt.cex}{double; cex for annotation text (default: 1.2)}

\item{my.title}{character; title of the figure (default: "")}
}
\description{
plot gene expression density
}
\details{
make density plot of genes. Note, density estimation from ggplot2 is different (and not so 'effective' as that from ks::kde). One of gene.id and gene.symbol must be provided.
}
