% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{ssc.reduceDim}
\alias{ssc.reduceDim}
\title{Reduce dimension by various methods}
\usage{
ssc.reduceDim(obj, assay.name = "exprs", method = "iCor",
  method.vgene = "sd", pca.npc = NULL, tSNE.usePCA = T,
  tSNE.perplexity = 30, zinbwave.K = 20, zinbwave.X = "~patient",
  autoTSNE = T, dim.name = NULL, iCor.niter = 1,
  iCor.method = "spearman", reuse = F, ncore = NULL, seed = NULL)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{method}{character; method to be used for dimension reduction, should be one of (pca, tsne, iCor, zinbwave). (default: "iCor")}

\item{method.vgene}{character; method to identify variable genes. (default: sd)}

\item{pca.npc}{integer; number of pc be used. Only for reduction method "pca". (default: NULL)}

\item{tSNE.usePCA}{logical; whether use PCA before tSNE. Only for reduction method "tsne". (default: T)}

\item{tSNE.perplexity}{logical; perplexity parameter. Used in all Rtsne() calling. (default: 30)}

\item{zinbwave.K}{integer, zinbwave parameter, number of latent variables. (default: 20)}

\item{zinbwave.X}{character, zinbwave parameter, cell-level covariates. (default: "~patient")}

\item{autoTSNE}{logical; Wheter generate automatically a tSNE map when reduction method is "pca" or "iCor". (default: T)}

\item{dim.name}{character; store the reduced data under the name in the obj's reducedDim SimpleList. If i
it is NULL, infer from method. (default: NULL)}

\item{iCor.niter}{integer; number of iteration of calculating the correlation. Used in reduction method "iCor". (default: 1)}

\item{iCor.method}{character; method to calculate correlation between samples,
should be one of "spearman" and "pearson". (default "spearman")}

\item{reuse}{logical; don't calculate if the query is already available. (default: F)}

\item{ncore}{integer; nuber of CPU cores to use. if NULL, automatically detect the number. (default: NULL)}

\item{seed}{integer; seed of random number generation. (default: NULL)}
}
\value{
an object of \code{SingleCellExperiment} class with reduced data added to the reducedDim slot.
}
\description{
Reduce dimension by various methods
}
\details{
If the reduction method is "pca", the function will call prcomp() and estimate the number of top PC should be used
in downstream analysis using and "elbow" based method, then the samples coordinates in the space spaned by the top PC would
stored in the reducedDim slot of the return value with the reducedDimName "pca".If autoTSNE is `true`, a tSNE map based on
the top PC will generated and stored in the reducedDim slot with reducedDimName "pca.tsne".

If the reduction method is "tsne", a tSNE map will generated and stored in the reducedDim slot with reducedDimName "tsne".
If tSNE.usePCA is `true`, maximum 50 top PC will be used, but no estimation based on "elbow" method performed.

If the reduction method is "iCor", the function will calculate correlation between samples for iCor.niter times.
If autoTSNE is `true`, it also generate automatically a tSNE map based on the correlations. These data would be stored in the
reducedDim slot with reducedDimName "iCor" and "iCor.tsne" respectively.
}
