% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.scale}
\alias{ssc.scale}
\title{scale the data}
\usage{
ssc.scale(
  obj,
  gene.id,
  gene.symbol,
  assay.name = "norm_exprs",
  adjB = NULL,
  do.scale = F
)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{gene.id}{should be in rownames(obj); genes to plot}

\item{gene.symbol}{should be in rowData(obj)[,"display.name"]; genes to plot}

\item{assay.name}{character; which assay (default: "exprs")}

\item{adjB}{character; batch column of the colData(obj). (default: NULL)}

\item{do.scale}{logical; whether scale the data. (default: F)}
}
\description{
scale the data
}
\details{
scale the data specified in assay.name, then store the scaled data to ${assay.name}.scale. One of gene.id and gene.symbol must be provided.
}
