% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.toLongTable}
\alias{ssc.toLongTable}
\title{convert assay data to long format}
\usage{
ssc.toLongTable(obj, gene.id, gene.symbol, assay.name = NULL, col.idx = NULL)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{gene.id}{should be in rownames(obj); genes to plot}

\item{gene.symbol}{should be in rowData(obj)[,"display.name"]; genes to plot}

\item{assay.name}{character; which assay. NULL for all assays. (default: NULL)}

\item{col.idx}{character; output extra columns of the colData(obj). NULL for donnot output extra columns. (default: NULL)}
}
\value{
Returns a data.table
}
\description{
convert assay data to long format
}
\details{
convert assay data to table of long format. It can be specified which genes and assays will be contained in the long table.
}
