% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{ssc.variableGene}
\alias{ssc.variableGene}
\title{Identify variable genes}
\usage{
ssc.variableGene(obj, method = "sd", sd.n = 1500, mean.thre = 0.1,
  assay.name = "exprs", reuse = F, out.prefix = NULL)
}
\arguments{
\item{obj}{object of SingleCellExperiment}

\item{method}{method to be used, can be one of "sd", mean.sd, trendVar. (default: "sd")}

\item{sd.n}{top number of genes (default 1500)}

\item{mean.thre}{numeric; threshold for mean, used in trendVar method (default 0.1)}

\item{assay.name}{which assay to be used (default "exprs")}

\item{reuse}{logical; don't calculate if the query is already available. (default: F)}

\item{out.prefix}{character; if not NULL, output prefix. (default: F)}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
Identify variable genes which will be used in downstream analysis. Multiple methods are available.
}
\details{
Method "sd": calculate the standard deviation of each gene and sort decreasingly, the top `sd.n` genes are the
variable genes. Method "trendVar": fit the trend between variance and mean, and decompose each gene's variance into
'tech' part(fitted value) and 'bio' part (residual value), then select genes according FDR and mean threshold.
}
