% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{binarizeExp}
\alias{binarizeExp}
\title{binarize the expression value using the distribution}
\usage{
binarizeExp(x, out.prefix = NULL, G = NULL, topNAsHi = 1,
  e.TH = NULL, e.name = "Exp", verbose = F, draw.CI = T,
  zero.as.low = T, run.extremevalue = F, zero.notUsed = F,
  my.seed = 9997, ...)
}
\arguments{
\item{x}{object; vector}

\item{out.prefix}{character; output prefix [default: NULL]}

\item{G}{integer; number of components [default: NULL]}

\item{topNAsHi}{integer; treat this number of top components as high;if it's zero, components with mean larger than 0 are high [default: 1]}

\item{e.TH}{double; for plot purpose: comparing the components with the user defined threhold [default: NULL]}

\item{e.name}{character; name of the expression metric [default: "Exp"]}

\item{verbose}{logical; [default: F]}

\item{draw.CI}{logical; [default: T]}

\item{zero.as.low}{logical; [default: T]}

\item{run.extremevalue}{logical; [default: F]}

\item{zero.notUsed}{logical; [default: F]}

\item{my.seed}{integer; [default: 9997]}

\item{...}{parameters passed to plot.densityMclust}
}
\description{
binarize the expression value using the distribution
}
\details{
use mixture model to classify each data point. If G is null, the largest component is corresponding to binarized expression 1, and other components are corresponding to binarized expressed 0.
}
