% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{ggGeneOnTSNE}
\alias{ggGeneOnTSNE}
\title{Plot gene expression on tSNE map}
\usage{
ggGeneOnTSNE(Y, dat.map, gene.to.show, out.prefix = NULL, p.ncol = 3,
  xlim = NULL, ylim = NULL, size = NULL, width = 9, height = 8)
}
\arguments{
\item{Y}{matrix or data.frame; Gene expression data, rownames shoud be gene id, colnames
should be sample id}

\item{dat.map}{data.frame; tSNE map, must be two columns data.frame and rownames should be sample id}

\item{gene.to.show}{character; gene id to be showd on the tSNE map}

\item{out.prefix}{character; output prefix (default: NULL)}

\item{p.ncol}{integer; number of columns in the plot's layout (default: 3)}

\item{xlim}{integer or NULL; only draw points lie in the ragne specified by xlim and ylim (default NULL)}

\item{ylim}{integer or NULL; only draw points lie in the ragne specified by xlim and ylim (default NULL)}

\item{size}{double; points' size. If NULL, infer from number of points (default NULL)}

\item{width}{numeric; width of the plot (default: 9)}

\item{height}{numeric; height of the plot (default: 8)}
}
\value{
a ggplot object
}
\description{
Plot gene expression on tSNE map
}
\details{
For genes contained in both `Y` and `gene.to.show`, show their expression on the tSNE
map provided as `dat.map`. One point in the map represent a cell; cells with higher expression
also have darker color.
}
