% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate.by.avg}
\alias{integrate.by.avg}
\title{integrate dataset by calculating the average expression and then clustering the pseudo bulk data}
\usage{
integrate.by.avg(
  sce.list,
  out.prefix,
  assay.name = "exprs",
  is.avg = FALSE,
  ncores = 6,
  use.deg = TRUE,
  gene.de.list = NULL,
  sort.by = "F.rank.median",
  avg.by = "majorCluster",
  n.downsample = NULL,
  n.pc = 15,
  do.clustering = T,
  de.stat = "t",
  de.thres = 1,
  do.scale = F,
  par.clust = list(method = "SNN", SNN.k = 3, SNN.method = "leiden",
    resolution_parameter = 2.2),
  topGene.lo = -1.5,
  topGene.hi = 1.5,
  topGene.step = 1,
  myseed = 9997,
  verbose = F,
  method.avg = "zscore",
  ...
)
}
\arguments{
\item{sce.list}{list; list of object of \code{singleCellExperiment} class}

\item{out.prefix}{character; output prefix.}

\item{assay.name}{character; which assay (default: "exprs")}

\item{is.avg}{logical; whether the objects in sce.list are average expression (default: FALSE)}

\item{ncores}{integer; number of cores to use (default: 6)}

\item{use.deg}{logical; whether use only the differentially expressed genes (default: FALSE)}

\item{gene.de.list}{list; if not NULL, each element is a data.frame with "geneID" column (default: NULL)}

\item{sort.by}{character; by which to sort genes and the top ones will be used for clustering. One of "F.rank.median", and "occurence" (default: "F.rank.median")}

\item{avg.by}{character; calculate the average expression of cells group by the specifid column (default: "majorCluster")}

\item{n.downsample}{integer; number of cells in each cluster to downsample to (default: NULL)}

\item{n.pc}{integer; number of pc ot use (default: 15)}

\item{do.clustering}{logical; wheter perform PCA/clustering (default: TRUE)}

\item{de.stat}{character; column in gene.de.file (default: "t")}

\item{de.thres}{double; used for selecting genes. can be larger than 1(top number of genes) or less than 1 (F.rank threshold) (default: 2000)}

\item{do.scale}{logical; scale the summarized vectors (default: false)}

\item{topGene.lo}{double; for top gene heatmap.(default: -1.5)}

\item{topGene.hi}{double; for top gene heatmap.(default: 1.5)}

\item{topGene.step}{double; for top gene heatmap.(default: 1)}

\item{myseed}{integer; seed for random number generation.(default: 9997)}

\item{verbose}{logical; verbose level.(default: FALSE)}

\item{method.avg}{character; method of calculate the average expression. Passed to `avg` of `ssc.average.cell`.(default: "zscore")}

\item{...}{parameters passed to ssc.clusterMarkerGene}
}
\description{
integrate dataset by calculating the average expression and then clustering the pseudo bulk data
}
\details{
method to calculate the average expression can be one of "mean", "zscore"
}
