% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{plot.density2D}
\alias{plot.density2D}
\title{Wrap for plotting 2D density}
\usage{
plot.density2D(x)
}
\arguments{
\item{x}{matrix or data.frame; map data, row for sample, column for dimension}
}
\description{
Wrap for plotting 2D density
}
\details{
use ks::kde for density estimation
}
