% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDistFromCellInfoTable}
\alias{plotDistFromCellInfoTable}
\title{plot distribution using cell info table}
\usage{
plotDistFromCellInfoTable(
  obj,
  out.prefix,
  plot.type = "barplot",
  facet.ncol = 3,
  plot.width = 10,
  plot.height = 5,
  test.method = "fisher.test",
  group.filter = NULL,
  sort.freq = F,
  bar.position = position_dodge2(),
  cmp.var = "Species",
  group.var = "ClusterID",
  donor.var = "donor",
  verbose = F,
  ...
)
}
\arguments{
\item{obj}{object; can be class of Seurat, SingleCellExperiment or data.frame;}

\item{out.prefix}{character; output prefix.}

\item{plot.type}{character; (default: "barplot")}

\item{facet.ncol}{integer; (default: 3)}

\item{plot.width}{integer; (default: 10)}

\item{plot.height}{integer; (default: 5)}

\item{test.method}{character; (default: "fisher.test")}

\item{group.filter}{character; (default: NULL)}

\item{sort.freq}{logical; (default: FALSE)}

\item{bar.position}{; passed to ggbarplot's position (default: position_dodge2())}

\item{cmp.var}{character; (default: "Species")}

\item{group.var}{character; (default: "ClusterID")}

\item{donor.var}{character; (default: "donor")}

\item{verbose}{logical; (default: FALSE)}

\item{...}{parameter passed to ggpubr}
}
\description{
plot distribution using cell info table
}
\details{
plot distribution
}
