% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.tSNE}
\alias{run.tSNE}
\title{Wraper for running Rtsne}
\usage{
run.tSNE(
  idata,
  tSNE.usePCA = T,
  tSNE.perplexity = 30,
  method = "Rtsne",
  n.cores = NULL,
  out.prefix = NULL,
  ...
)
}
\arguments{
\item{idata}{matrix; expression data with sample id in rows and variables in columns}

\item{tSNE.usePCA}{whether perform PCA before tSNE (default: T)}

\item{tSNE.perplexity}{perplexity parameter of tSNE (default: 30)}

\item{method}{method to be used. one of "Rtsne" and "FIt-SNE" (default: "Rtsne")}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{out.prefix}{character; output prefix (default: NULL)}

\item{...}{parameters passed to Rtsne or fftRtsne}
}
\value{
If successful same as the return value of Rtsne(); otherwise NULL
}
\description{
Wraper for running Rtsne
}
