% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.average.cell}
\alias{ssc.average.cell}
\title{calcualte the average expression of cells of each group}
\usage{
ssc.average.cell(obj, assay.name = "exprs", gene = NULL,
  column = "majorCluster", ncell.downsample = NULL, avg = "mean",
  ret.type = "data.melt")
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{gene}{character; only consider the specified gnees (default: NULL)}

\item{column}{character; columns in colData(obj) to be averaged. (default: "majorCluster")}

\item{ncell.downsample}{integer; for each group, number of cells downsample to. (default: NULL)}

\item{avg}{character; average method. can be one of "mean", "diff", "zscore" . (default: "mean")}

\item{ret.type}{character; return type. can be one of "data.melt", "data.cast", "data.mtx". (default: "data.melt")}
}
\description{
calcualte the average expression of cells of each group
}
\details{
multiple average methods are implemented
}
