% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.clust}
\alias{ssc.clust}
\title{Perform clustering using reduced data}
\usage{
ssc.clust(obj, assay.name = "exprs", method.reduction = "iCor",
  method = "kmeans", k.batch = 2:6, method.vgene = "sd", SNN.k = 10,
  SNN.method = "eigen", SC3.biology = T, SC3.markerplot.width = 15,
  dpclust.rho = NULL, dpclust.delta = NULL, parlist = NULL,
  out.prefix = NULL, seed = NULL, ncore = NULL)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{method.reduction}{character; which dimention reduction method to be used, should be one of
"iCor", "pca" and "none". (default: "iCor")}

\item{method}{character; clustering method to be used, should be one of "kmeans", "hclust", "SNN", "dpclust", "adpclust" and "SC3". (default: "kmeans")}

\item{k.batch}{integer; number of clusters to be evaluated. (default: 2:6)}

\item{method.vgene}{character; variable gene identification method used. (default: "sd")}

\item{SNN.k}{integer; number of shared NN. (default: 10)}

\item{SNN.method}{character; cluster method applied on SNN， one of "greedy", "eigen", "infomap",
"prop", "louvain", "optimal", "spinglass", "walktrap", "betweenness". (default: "eigen")}

\item{SC3.biology}{logical, SC3 parameter, whether calcualte biology. (default: T)}

\item{SC3.markerplot.width}{integer, SC3 parameter, with of the marker plot (default: 15)}

\item{dpclust.rho}{numberic; cuttoff of rho, if it is NULL, infer frome the data (default: NULL)}

\item{dpclust.delta}{numberic; cuttoff of delta, if it is NULL, infer frome the data (default: NULL)}

\item{parlist}{list; if not NULL, use th parameters in it. (default: NULL)}

\item{out.prefix}{character; output prefix, if not NULL, some plots of intermediate result will be produced. (default: NULL)}

\item{seed}{integer; seed of random number generation. (default: NULL)}

\item{ncore}{integer; nuber of CPU cores to use. if NULL, automatically detect the number. (default: NULL)}
}
\value{
an object of \code{SingleCellExperiment} class with cluster labels added.
}
\description{
Perform clustering using reduced data
}
\details{
If no dimension reduction performed or method is "none", expression data of variable genes,
which can be speficed by method.vgene, will be used for clustering. Otherwise, the reduced data specified by
method.reduction will be used. The cluster label will stored in the colData of the object
of \code{singleCellExperiment} class, with colname in the format of \{method.reduction\}.\{method\}k\{k\}
where \{k\} get value(s) from k.batch.
}
