% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.clusterMarkerGene}
\alias{ssc.clusterMarkerGene}
\title{identify marker genes of each cluster}
\usage{
ssc.clusterMarkerGene(obj, assay.name = "exprs", group.var = "majorCluster",
  assay.bin = NULL, minCell = 5, out.prefix, n.cores = NULL,
  do.plot = T, F.FDR.THRESHOLD = 0.05, pairwise.P.THRESHOLD = 0.05,
  pairwise.FC.THRESHOLD = 1, verbose = F)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{p.ncol}{integer; number of columns in the figure layout. (default: 2)}

\item{exp.bin;}{numeric; binarized expression matrix, rows for genes and columns for samples. value 1 means expressed.}

\item{exp.norm;}{numeric; expression matrix, rows for genes and columns for samples. original version of exp.bin.}

\item{group;}{character; clusters of the samples belong to}
}
\description{
identify marker genes of each cluster
}
\details{
identify marker genes based on aov and AUC.
}
