% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.DEGene.limma}
\alias{ssc.DEGene.limma}
\title{identify differential genes of each cluster (comparing the cluster with all others), using limma}
\usage{
ssc.DEGene.limma(
  obj,
  assay.name = "exprs",
  ncell.downsample = NULL,
  group.var = "majorCluster",
  group.list = NULL,
  group.mode = "multi",
  batch = NULL,
  out.prefix = NULL,
  n.cores = NULL,
  do.plot = T,
  T.fdr = 0.01,
  T.logFC = 1,
  T.expr = 0.3,
  T.bin.useZ = T,
  verbose = 0,
  do.force = F,
  method = "limma"
)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{ncell.downsample}{integer; for each group, number of cells downsample to. (default: NULL)}

\item{group.var}{character; column in the colData(obj) used for grouping. (default: "majorCluster")}

\item{group.list}{character; DEG of groups to calculate. If NULL, all groups. (default: "NULL")}

\item{group.mode}{character; One of "multi", "multiAsTwo" (default: "multi")}

\item{batch}{character; covariate. (default: NULL)}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{n.cores}{integer; number of cores used, if NULL it will be determined automatically (default: NULL)}

\item{do.plot}{logical; whether plot. (default: TRUE)}

\item{T.fdr}{numeric; threshold of the adjusted p value of moderated t-test (default: 0.05)}

\item{T.logFC}{numeric; threshold of the absoute diff (default: 1)}

\item{T.expr}{numeric; threshold for binarizing exprs (default: 0.3)}

\item{T.bin.useZ}{logical; wheter use the z-score version of assay.namme for binarizing exprs (default: T)}

\item{verbose}{integer; verbose level. (default: 0)}

\item{do.force}{logical; . (default: FALSE)}

\item{method}{character; . (default: "limma")}
}
\description{
identify differential genes of each cluster (comparing the cluster with all others), using limma
}
\details{
identify differential genes using limma
}
