% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{ssc.plot.violin}
\alias{ssc.plot.violin}
\title{Plot violin}
\usage{
ssc.plot.violin(obj, assay.name = "exprs", gene = NULL, columns = NULL,
  group.var = "majorCluster", out.prefix = NULL, p.ncol = 1,
  base_aspect_ratio = 1.1)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{gene}{character; genes to be showed. (default: NULL)}

\item{columns}{character; columns in colData(obj) to be showd. (default: NULL)}

\item{group.var}{character; column in the colData(obj) used for grouping. (default: "majorCluster")}

\item{out.prefix}{character; output prefix. (default: NULL)}

\item{p.ncol}{integer; number of columns in the figure layout. (default: 3)}

\item{base_aspect_ratio}{numeric; base_aspect_ratio, used for plotting metadata. (default 1.1)}
}
\description{
Plot violin
}
\details{
If `gene` is not NULL, violin of the genes' expression will be plot; if columns in not
NULL, colData of obj with names in `columns` will be plot in violin.
}
