% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.run}
\alias{ssc.run}
\title{Wrapper for running all the pipeline}
\usage{
ssc.run(obj, assay.name = "exprs", method.vgene = "sd", sd.n = 1500,
  mean.thre = 0.1, fdr.thre = 0.001, var.block = NULL,
  method.reduction = "iCor", method.clust = "kmeans",
  method.classify = "knn", pca.npc = NULL, iCor.niter = 1,
  iCor.method = "spearman", zinbwave.K = 20, zinbwave.X = "~patient",
  subsampling = F, sub.frac = 0.4, sub.use.proj = T, sub.vis.proj = F,
  k.batch = 2:6, refineGene = F, de.n = 1500, HSD.FC.THRESHOLD = 1,
  nIter = 1, out.prefix = NULL, parfile = NULL, reuse = F,
  ncore = NULL, seed = NULL, do.DE = F, ...)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{method.vgene}{character; variable gene identification method used. (default: "sd")}

\item{sd.n}{integer; top number of genes as variable genes (default 1500)}

\item{mean.thre}{numeric; threshold for mean, used in trendVar method (default 0.1)}

\item{fdr.thre}{numeric; threshold for fdr, used in trendVar method (default 0.001)}

\item{var.block}{character; specify the uninteresting factors by formula. E.g. "~patient".
used in trendVar method (default NULL)}

\item{method.reduction}{character; which dimention reduction method to be used, should be one of
"iCor", "pca", "zinbwave" and "none". (default: "iCor")}

\item{method.clust}{character; clustering method to be used, should be one of "kmeans", "hclust", "SNN", "adpclust" and "SC3. (default: "kmeans")}

\item{method.classify}{character; method used for classification, one of "knn" and "RF". (default: "knn")}

\item{pca.npc}{integer; number of pc be used. Only for reduction method "pca". (default: NULL)}

\item{iCor.niter}{integer; number of iteration of calculating the correlation. Used in reduction method "iCor". (default: 1)}

\item{iCor.method}{character; correlation method, one of "spearman", "pearson" (default: "spearman")}

\item{zinbwave.K}{integer, zinbwave parameter, number of latent variables. (default: 20)}

\item{zinbwave.X}{character, zinbwave parameter, cell-level covariates. (default: "~patient")}

\item{subsampling}{logical; whether cluster using the subsampling->cluster->classification method. (default: F)}

\item{sub.frac}{numeric; subsample to frac of original samples. (default: 0.4)}

\item{sub.use.proj}{logical; whether use the projected data for classification. (default: T)}

\item{sub.vis.proj}{logical; whether get low dimensional representation for visualization, only used in downsample mode. (default: F)}

\item{k.batch}{integer; number of clusters to be evaluated. (default: 2:6)}

\item{refineGene}{logical; whether perform second round demension reduction and clustering pipeline using the differential
genes found by the first round cluster result. (default: F)}

\item{de.n}{integer; number of differential genes used for refined geneset for another run of clustering (default 1500)}

\item{HSD.FC.THRESHOLD}{numeric; threshold for log2FoldChange, used in findDEGenesByAOV (default 1)}

\item{nIter}{integer; number of iterative clustering in sub-cluster. (default: 1)}

\item{out.prefix}{character; output prefix, if not NULL, some plots of intermediate result will be produced. (default: NULL)}

\item{parfile}{character; parameter files, if not NULL, will use the settings. must contain a list named
`parlist`. (default: NULL)}

\item{reuse}{logical; don't calculate if the query is already available. (default: F)}

\item{ncore}{integer; nuber of CPU cores to use. if NULL, automatically detect the number. (default: NULL)}

\item{seed}{integer; seed of random number generation. (default: NULL)}

\item{do.DE}{logical; perform DE analysis when clustering finished. (default: F)}

\item{...}{parameters pass to clustering methods}
}
\value{
an object of \code{SingleCellExperiment} class with cluster labels added.
}
\description{
Wrapper for running all the pipeline
}
\details{
run the pipeline of variable gene identification, dimension reduction, clustering.
}
\seealso{
\code{\link{ssc.variableGene}} for variable genes' identification, \code{\link{ssc.reduceDim}}
for dimension reduction, \code{\link{ssc.clust}} for clustering using all data
and \code{\link{ssc.clustSubsamplingClassification}} for clustering with subsampling.
}
