% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccClust.R
\name{ssc.run}
\alias{ssc.run}
\title{Wrapper for running all the pipeline}
\usage{
ssc.run(obj, assay.name = "exprs", method.vgene = "sd", sd.n = 1500,
  method.reduction = "iCor", method.clust = "kmeans",
  method.classify = "knn", pca.npc = NULL, iCor.niter = 1,
  subsampling = F, sub.frac = 0.4, sub.use.proj = T, k.batch = 2:6,
  refineGene = F, iterative = F)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{method.vgene}{character; variable gene identification method used. (default: "sd")}

\item{sd.n}{top number of genes (default 1500)}

\item{method.reduction}{character; which dimention reduction method to be used, should be one of
"iCor", "pca" and "none". (default: "iCor")}

\item{method.clust}{character; clustering method to be used, should be one of "kmeans" and "hclust". (default: "kmeans")}

\item{method.classify}{character; method used for classification, one of "knn" and "RF". (default: "knn")}

\item{pca.npc}{integer; number of pc be used. Only for reduction method "pca". (default: NULL)}

\item{iCor.niter}{integer; number of iteration of calculating the correlation. Used in reduction method "iCor". (default: 1)}

\item{subsampling}{logical; whether cluster using the subsampling->cluster->classification method. (default: F)}

\item{sub.frac}{numeric; subsample to frac of original samples. (default: 0.4)}

\item{sub.use.proj}{logical; whether use the projected data for classification. (default: T)}

\item{k.batch}{integer; number of clusters to be evaluated. (default: 2:6)}

\item{refineGene}{logical; whether perform second round demension reduction and clustering pipeline using the differential
genes found by the first round cluster result. (default: F)}

\item{iterative}{logical; whether perform iterative clustering in sub-cluster. (default: F)}
}
\value{
an object of \code{SingleCellExperiment} class with cluster labels added.
}
\description{
Wrapper for running all the pipeline
}
\details{
run the pipeline of variable gene identification, dimension reduction, clustering.
}
\seealso{
\code{\link{ssc.variableGene}} for variable genes' identification, \code{\link{ssc.reduceDim}}
for dimension reduction, \code{\link{ssc.clust}} for clustering using all data
and \code{\link{ssc.clustSubsamplingClassification}} for clustering with subsampling.
}
