% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscClust.R
\name{ssc.variableGene}
\alias{ssc.variableGene}
\title{Identify variable genes}
\usage{
ssc.variableGene(obj, method = "HVG.sd", sd.n = 1500,
  mean.thre = 0.1, fdr.thre = 0.001, assay.name = "exprs",
  var.block = NULL, reuse = F, out.prefix = NULL)
}
\arguments{
\item{obj}{object of SingleCellExperiment}

\item{method}{method to be used, can be one of "HVG.sd", "HVG.mean.sd", "HVG.trendVar". (default: "HVG.sd")}

\item{sd.n}{top number of genes (default 1500)}

\item{mean.thre}{numeric; threshold for mean, used in trendVar method (default 0.1)}

\item{fdr.thre}{numeric; threshold for fdr, used in trendVar method (default 0.001)}

\item{assay.name}{which assay to be used (default "exprs")}

\item{var.block}{character; specify the uninteresting factors by formula. E.g. "~patient" (default NULL)}

\item{reuse}{logical; don't calculate if the query is already available. (default: F)}

\item{out.prefix}{character; if not NULL, output prefix. (default: F)}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
Identify variable genes which will be used in downstream analysis. Multiple methods are available.
}
\details{
Method "sd": calculate the standard deviation of each gene and sort decreasingly, the top `sd.n` genes are the
variable genes. Method "trendVar": fit the trend between variance and mean, and decompose each gene's variance into
'tech' part(fitted value) and 'bio' part (residual value), then select genes according FDR and mean threshold. Note,
when using "trendVar", will use expression data stored in "norm_exprs" slot of `obj`, no matter what `assay.name` is.
}
